refactor: completely remove references to wot-servient and rename it to node-wot
