+++ packages/cli/import-json-to-ts.js
@@ -1,7 +1,7 @@
 const { readFileSync, writeFileSync } = require("fs");
 const { existsSync, mkdirSync } = require("fs");
 
-const schema = readFileSync("./src/wot-servient-schema.conf.json", "utf8");
+const schema = readFileSync("./src/node-wot-schema.conf.json", "utf8");
 const package = readFileSync("./package.json", "utf8");
 const { version } = JSON.parse(package);
 
@@ -11,7 +11,7 @@ if (!existsSync(generatedDir)) {
 }
 
 writeFileSync(
-    "./src/generated/wot-servient-schema.conf.ts",
+    "./src/generated/node-wot-schema.conf.ts",
     `const schema = ${schema.trimEnd()} as const \nexport default schema;`
 );
 writeFileSync("./src/generated/version.ts", `const version = "${version}" as const \nexport default version;`);
