+++ packages/cli/src/node-wot-schema.conf.json
